agrego primeros DTOs, agrego un mapper para champion y primer endpoint de prueba

+++ src/main/java/com/repository/ChampionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChampionRepository extends JpaRepository<Champion,Integer> {
+public interface ChampionRepository extends JpaRepository<Champion,Long> {
 }
